+++ R/plotReadsPerCell-methods.R
@@ -1,6 +1,5 @@
 #' @name plotReadsPerCell
 #' @author Michael Steinbaugh, Rory Kirchner
-#' @include globals.R
 #' @inherit bioverbs::plotReadsPerCell
 #' @note Updated 2019-08-12.
 #'